+++ packages/tools/pie-icons-react/icons/FlagPoland.tsx
@@ -1,0 +1,6 @@
+import * as React from "react";
+import { SVGProps } from "react";
+
+const FlagPoland = (props: SVGProps<SVGSVGElement>) => <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 14 14" className="c-pieIcon c-pieIcon--flagPoland" width="1em" height="1em" role="img" {...props}><path d="M7 14C10.866 14 14 10.866 14 7C14 3.13401 10.866 0 7 0C3.13401 0 0 3.13401 0 7C0 10.866 3.13401 14 7 14Z" fill="#EEEEEE" /><path d="M14 7C14 8.85652 13.2625 10.637 11.9497 11.9497C10.637 13.2625 8.85652 14 7 14C5.14348 14 3.36301 13.2625 2.05025 11.9497C0.737498 10.637 0 8.85652 0 7" fill="#D80027" /></svg>;
+
+export default FlagPoland;