+++ packages/tools/pie-icons-react/icons/FullscreenExit.tsx
@@ -1,0 +1,6 @@
+import * as React from "react";
+import { SVGProps } from "react";
+
+const FullscreenExit = (props: SVGProps<SVGSVGElement>) => <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 28 28" className="c-pieIcon c-pieIcon--fullscreenExit" width="1em" height="1em" role="img" {...props}><path d="M19.25 2.1875L17.5 1.3125V10.5H26.632L25.6461 8.75H19.25V2.1875Z" /><path d="M2.35385 19.25L1.36794 17.5H10.5V26.6875L8.74998 25.8125V19.25H2.35385Z" /><path d="M26.6875 17.5L25.8125 19.25H19.25V25.6461L17.5 26.632V17.5H26.6875Z" /><path d="M8.75 2.35385L10.5 1.36794V10.5H1.3125L2.1875 8.74998H8.75V2.35385Z" /></svg>;
+
+export default FullscreenExit;