+++ packages/tools/pie-icons-react/icons/GeolocationFilledSmall.tsx
@@ -1,0 +1,6 @@
+import * as React from "react";
+import { SVGProps } from "react";
+
+const GeolocationFilledSmall = (props: SVGProps<SVGSVGElement>) => <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 14 14" className="c-pieIcon c-pieIcon--geolocationFilledSmall" width="1em" height="1em" role="img" {...props}><path d="M0.534668 5.54755V7.00005L2.66967 7.5338C3.59936 7.76292 4.44868 8.24192 5.12574 8.91898C5.8028 9.59604 6.2818 10.4454 6.51092 11.3751L7.04467 13.5101H8.45342L12.4172 1.5838L0.534668 5.54755Z" /></svg>;
+
+export default GeolocationFilledSmall;