+++ packages/tools/pie-icons-react/icons/PlusSmall.tsx
@@ -1,0 +1,6 @@
+import * as React from "react";
+import { SVGProps } from "react";
+
+const PlusSmall = (props: SVGProps<SVGSVGElement>) => <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 14 14" className="c-pieIcon c-pieIcon--plusSmall" width="1em" height="1em" role="img" {...props}><path d="M13.125 6.34375H7.65625V0.875H6.34375V6.34375H0.875V7.65625H6.34375V13.125H7.65625V7.65625H13.125V6.34375Z" /></svg>;
+
+export default PlusSmall;