+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v1.17.0
+------------------------------
+*October 22, 2022*
+
+### Added
+- `pie-icons-react` package to `/tools`.
+- add `/esm` folders to .gitignore
+
 
 v1.16.1
 ------------------------------